@@ -1,16 +1,10 @@
 ﻿using Apps.ContentQuo.Dtos;
 using Blackbird.Applications.Sdk.Common;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Apps.ContentQuo.Models.Responses
+namespace Apps.ContentQuo.Models.Responses;
+
+public class ListBilingualFilesResponse
 {
-    public class ListBilingualFilesResponse
-    {
-        [Display("Bilingual files")]
-        public List<BilingualFileDto> BilingualFiles { get; set; }
-    }
+    [Display("Bilingual files")]
+    public List<BilingualFileDto> BilingualFiles { get; set; }
 }
