@@ -1,14 +1,8 @@
 ﻿using Apps.ContentQuo.Dtos;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Apps.ContentQuo.Models.Responses
+namespace Apps.ContentQuo.Models.Responses;
+
+public class ListEvaluationIssuesResponse
 {
-    public class ListEvaluationIssuesResponse
-    {
-        public List<IssueDto> Issues { get; set; }
-    }
+    public List<IssueDto> Issues { get; set; }
 }
